flask_app/requirements.txt
opencv-python==4.6.0.66
onnxruntime==1.13.1
tf2onnx==1.13.0
tensorflow==2.10.0
numpy==1.23.5
pillow==9.2.0
keras==2.10.0
flask==2.2.2
scipy==1.9.3